pyproject.toml
[project]
name = "mem0-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dotenv>=0.9.9",
    "langchain-neo4j>=0.4.0",
    "mem0ai>=0.1.94",
    "rank-bm25>=0.2.2",
]
